chore: remove useless component types (#5219)
