--- controllers/apps/components/rsm_workload.go
@@ -1,47 +1,0 @@
-/*
-Copyright (C) 2022-2023 ApeCloud Co., Ltd
-
-This file is part of KubeBlocks project
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU Affero General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Affero General Public License for more details.
-
-You should have received a copy of the GNU Affero General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-package components
-
-import (
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	"github.com/apecloud/kubeblocks/internal/controller/factory"
-)
-
-type rsmComponentWorkloadBuilder struct {
-	componentWorkloadBuilderBase
-}
-
-var _ componentWorkloadBuilder = &rsmComponentWorkloadBuilder{}
-
-func (b *rsmComponentWorkloadBuilder) BuildWorkload() componentWorkloadBuilder {
-	buildfn := func() ([]client.Object, error) {
-		component := b.Comp.GetSynthesizedComponent()
-		obj, err := factory.BuildRSM(b.ReqCtx, b.Comp.GetCluster(), component, b.EnvConfig.Name)
-		if err != nil {
-			return nil, err
-		}
-
-		b.Workload = obj
-
-		return nil, nil // don't return sts here
-	}
-	return b.BuildWrapper(buildfn)
-}
